% $Id: rs00-RPPASPACE-package.Rd

\name{RPPASPACE-package}
\alias{RPPASPACE-package}
\docType{package}
\title{Reverse phase protein lysate array analysis}
\description{
  A package for analyzing reverse phase protein lysate arrays (RPPA).
}
\details{
  \tabular{ll}{
    Package: \tab RPPASPACE\cr
    Type:    \tab Package\cr
    Version: \tab \Sexpr[results=text]{packageDescription("RPPASPACE")$Version}\cr
    Phase:   \tab \Sexpr[results=text]{packageDescription("RPPASPACE")$Phase}\cr
    Date:    \tab \Sexpr[results=text]{packageDescription("RPPASPACE")$Date}\cr
    License: \tab \Sexpr[results=text]{packageDescription("RPPASPACE")$License}\cr
  }

  For a complete list of functions, use \code{library(help="RPPASPACE")}.\cr
  For a high-level summary of the changes for each revision, use\cr
    \code{file.show(system.file("NEWS", package="RPPASPACE"))}.
}
\author{
  Kevin R. Coombes \email{coombes.3@osu.edu},
  P. Roebuck \email{paul_roebuck@comcast.net},
  James M. Melott \email{jmmelott@mdanderson.org}
}
\keyword{package}

